refactor: Reorganize folders by feature and add initial performance optimizations
This commit introduces a significant restructuring of the project's folders for better maintainability and applies initial performance enhancements for the mind map component.

1.  **Folder Structure Reorganization:**
    *   I restructured the `src/` directory to group files by features (e.g., `src/features/mind-map/`, `src/features/sidebar/`, `src/features/quiz/`, `src/features/ai/`) and core application concerns (`src/app/`).
    *   Shared UI components remain in `src/components/ui/`, general utilities in `src/lib/`, and contexts/hooks/pages in their respective top-level directories.
    *   I updated all import paths across the application to reflect the new file locations. This change improves modularity and makes it easier to locate feature-specific code.

2.  **Performance Optimizations for Large Mind Maps (Initial Steps):**
    *   I enabled the `onlyRenderVisibleElements={true}` prop on the `<ReactFlow />` component within `MindMap.tsx`. This instructs React Flow to only render nodes and edges currently within or near the viewport, which can significantly improve rendering performance for large graphs.
    *   I reviewed and confirmed existing memoization strategies (`React.memo` for `CustomNode`, `useCallback` and `useMemo` in `MindMapProvider` and `MindMap.tsx`) to ensure efficient rendering and prevent unnecessary re-renders.
    *   I updated the dependency array for the `useMemo` hook creating the `MindMapContext` value for correctness.

These changes lay a better foundation for future development and aim to provide a smoother experience when dealing with more complex mind maps.
