+++ src/features/mind-map/components/CustomNode.tsx
