+++ src/features/mind-map/config/initial-elements.ts
