+++ src/features/sidebar/components/AiSettingsDialog.tsx
