+++ src/features/sidebar/components/FileControls.tsx
